api/core/modelo_pixel/requirements.txt
# Núcleo IA (Orden crítico)
torch>=2.3.0
torchvision>=0.18.0

# Librerías para Florence-2
transformers>=4.41.2
pillow
einops
timm
accelerate
scipy
# flash_attn  